requirements.txt
google-generativeai
python-dotenv
flask
werkzeug
PyPDF2 